feat: add support for more diagrams
